Removing additional fields to bitcoin.js
Adding non-standard fields to to third-party objects can cause confusion and decrease maintanability in long-term. (It's not clear when exactly are they present and when exactly aren't.)

I removed the fields from the declaration file and moved them to a separate fields in TransactionInfo. Now it's 100% sure that they will be present in TransactionInfo, since they are created in the constructor and not in fromJSON - before this change, the fields wouldn't be present when created from BitcoreBlockchain class. (Exactly the kind of error that type systems prevent :) )

I have *not* tested if it works.