Correct bug with unconfirmed txs
The forceAddTransaction was added, so that immediately after transaction send, user see the TX in his list, even when the backend didn't yet send the TX. This sometimes happened when the backend was too busy, so users didn't see their own transactions and had wrong UTXO sets.

The original intent was to add the TX to the array, but remove it once it has been seen in the stream from the backend.

However, there has been an error in the code and the TX did not get removed and stayed there until window reload. That means that the unconfirmed transaction always "overloaded" the confirmed one, so transactions stayed unconfirmed.

That was one problem - however, there was an additional consequence introduced later. In the recent update, hd-wallet removes unconfirmed transactions that haven't been seen in either recent update or in mempool. This would normally work - however, it broke when transaction was wrongly saved as unconfirmed, but it was not in the forceAddedTransaction array anymore (because that is only saved in memory)

This fixes both bugs - transactions are now displayed as confirmed when they are confirmed, and they are not removed on reload.