fix(tor): Allow user go back to default address by clearing the input
When the address we cleared in UI the proxy was actually configured to
route traffic to an empty address. This wasn't apparent to the user
because the placeholder was shown in that case.

When user sets the address to empty string set it to the default
instead. At the same time use the default as placeholder to indicate
this to the user.