chore: Fail early on missing static files
If some of the explicitly required static files are missing the
`copy-static-files` job logs that as error but doesn't fail.

Make it fail loudly to catch missing files as early as possible.
Extend the global README with command to generate message system config
signature which is now required for `copy-static-files` to succeed.