fix(guide): #3881 Fix Suite layout for various widths and guide states
The first version of Guide was implemented in a way that opening caused
the accounts menu to hide.

Introduce new layout option to handle all combinations of widths and
guide states.

When the viewport width is XLARGE (+1200px) display menu, body and guide
in three columns along each other.

When the viewport is smaller than XLARGE but not larger than mobile
(768-1200px) display menu and body in two columns unless the guide is
open. In that case display body and guide in two columns and inline the
menu on top of body.

If the viewport is below the mobile threshold preserve the current
state. That is inline the menu on top of body and hide guide completely.