fix: Category Select state lifecycle
The stateful `category` variable was initialized to a different value
('dashboard') than what has been set as default value of the Select
component. This causes a submitting potentially invalid data and
prevents validation of non-empty values.

Use the same value for initialization of both the Select components and
the stateful variable. Shuffle the code a bit to satisfy the new order
of usages.