fix(suite): fix suite stucked in bootloader mode
This removes race condition (wrong order of two subsequent @suite/select-device actions) when there is a need to reset wallet params.

Steps to reproduce was
1. load /accounts#/btc/0 without device connected
2. connect device in bootloader and then
3. disconnect it.

Awaiting on dispatching actions doesn't guarantee waiting for the action to be processed anyway if the action creator is not returning promise.

So the solution here is to make dispatching @suite/select-device synchronous.