fix(suite): disable Solana websockets and account subscriptions
It seems that there are some issues with the websockets connection which are hard to identify and find. They seem to be happening quite randomly or in very specific cases i.e. they can't be reproduced and thus are very hard to fix. For now we decided to disable the Solana SDK websockets connection and thus we also disabled account subscriptions (which were supposedly causing the issues).

(cherry picked from commit 3c6f925e4c56a0ae7f73ae75d7517818ddc23895)