fix(blockchain-link): use constant and increased compute unit limit for Solana transactions
In an ideal world over-reserving compute units would not result in faster transaction confirmation but if we optimize the compute unit limit to match the amount of CUs the transaction requires (+ a 20% buffer) the transactions simply do not go through in time.

Perhaps this could be revisited at some point in the future e.g. when [this PR](https://github.com/anza-xyz/agave/pull/187) is released and validators update.